pyproject.toml
[project]
name = "usda-vision-cameras"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "imageio>=2.37.0",
    "matplotlib>=3.10.3",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "paho-mqtt>=2.1.0",
    "pillow>=11.3.0",
    "tqdm>=4.67.1",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "websockets>=12.0",
    "requests>=2.31.0",
    "pytz>=2023.3",
    "ipykernel>=6.30.0",
    "httpx>=0.28.1",
    "aiofiles>=24.1.0",
]
